// source/frontend/src/models/api/generated/ApiGen_CodeTypes_DispositionFileStatusTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_DispositionFileStatusTypes {
  ACTIVE = 'ACTIVE',
  ARCHIVED = 'ARCHIVED',
  CANCELLED = 'CANCELLED',
  COMPLETE = 'COMPLETE',
  DRAFT = 'DRAFT',
  HOLD = 'HOLD',
}